+++ visa_manager_frontend/jest.config.js
@@ -22,7 +22,16 @@ module.exports = {
   coverageDirectory: 'coverage',
   coverageReporters: ['text', 'lcov', 'html'],
   moduleNameMapper: {
-    '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': 'identity-obj-proxy',
+    '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/__mocks__/assets.js',
+    '^@components/(.*)$': '<rootDir>/src/components/$1',
+    '^@context/(.*)$': '<rootDir>/src/context/$1',
+    '^@hooks/(.*)$': '<rootDir>/src/hooks/$1',
+    '^@navigation/(.*)$': '<rootDir>/src/navigation/$1',
+    '^@screens/(.*)$': '<rootDir>/src/screens/$1',
+    '^@services/(.*)$': '<rootDir>/src/services/$1',
+    '^@styles/(.*)$': '<rootDir>/src/styles/$1',
+    '^@types/(.*)$': '<rootDir>/src/types/$1',
+    '^@utils/(.*)$': '<rootDir>/src/utils/$1',
   },
   transformIgnorePatterns: [
     'node_modules/(?!(react-native|@react-native|react-native-paper|react-native-vector-icons|@react-navigation)/)',